+++ test/fixtures/example/language.js
@@ -1,0 +1,35 @@
+// Test for programming language: JavaScript
+
+/**
+ * @api {get} /language/javascript JavaScript
+ * @apiName GetLanguageJavaScript
+ * @apiGroup Language
+ * @apiVersion 0.4.0
+ * @apiDescription Test for JavaScript Comment-Syntax.
+ */
+
+/**
+ * @api {get} /language/javascript/indented1 JavaScript indented 1
+ * @apiName GetLanguageJavaScriptIndented1
+ * @apiGroup Language
+ * @apiVersion 0.4.0
+ * @apiExample Test for indented comment.
+ * This is example line 2.
+ * This is example line 3.
+ * 	    Line 4 indented (with tab at beginning).
+ *     Line 5 indented.
+ * This is example line 6.
+ */
+
+/**
+@api {get} /language/javascript/indented2 JavaScript indented 2
+@apiName GetLanguageJavaScriptIndented2
+@apiGroup Language
+@apiVersion 0.4.0
+@apiExample Test for indented comment.
+This is example line 2.
+This is example line 3.
+	    Line 4 indented (with tab at beginning).
+    Line 5 indented.
+This is example line 6.
+*/